Add support for SCP 079 generators to ProjectMER
Added Support for Generators.
Properties:
Permissions - The required permissions
IsUnlocked - If the Generator is unlocked
TotalActivationTime - The total activation time of the generator Engaged - Whether or not the generator is engaged.
Engaged - If the Generator Is Engaged

I somehow messed up my old github repository and I don't feel like fixing it so here is a new PR.